pyproject.toml
[tool.poetry]
name = "planning-through-contact"
version = "0.1.0"
description = ""
authors = ["Bernhard Paus Græsdal <bernhardpg@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.10, <3.11"
matplotlib = "^3.6.1"
numpy = "^1.23.3"
pdbpp = "^0.10.3"
tqdm = "^4.64.1"
pydot = "^1.4.2"
tk = "^0.1.0"
drake = "1.27"
pytest = "^7.4.0"
ffmpeg = "^1.4"
ffmpeg-python = "^0.2.0"
pygame = "^2.5.1"
opencv-python = "^4.9.0.80"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
zarr = "^2.16.1"

# Diffusion Policy dependencies
# diffusion-policy = {path = "/home/adam/workspace/gcs-diffusion/dist/diffusion_policy-0.0.0-py3-none-any.whl"}
dill = "^0.3.8"
torch = "^2.2.0"
diffusers = "0.11.1"
robomimic = "0.3.0"
einops = "0.4.1"
numba = "^0.59.0"
lxml = "^5.2.1"
manipulation = "2024.3.28"
pyvirtualdisplay = "^3.0"
scipy = "^1.13.0"
wandb = "^0.17.0"
huggingface-hub = ">=0.10.0,<0.25.3"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"
pre-commit = "^3.4.0"
black = "^23.12.1"
isort = "^5.12.0"
poetry-pre-commit-plugin = "^0.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
# PyRight wrongly reports that several Drake modules are missing (could also be an issue with Drake)
reportMissingImports = false

[tool.isort]
profile = "black" # Make isort and Black agree on sorting